+++ PitchDetection/src/PitchDetector/PitchDetector.h
@@ -20,10 +20,10 @@ typedef enum Note_ {
 } Note;
 
 typedef struct PitchInfo_ {
-	double freq;
+	float freq;
 	Note note;
 	const char* noteStr;
-	uint32_t octave;
+	uint8_t octave;
 } PitchInfo;
 
 class PitchDetector
@@ -32,19 +32,19 @@ public:
 	PitchDetector(int samplingRate, int samplingSize);
 	virtual ~PitchDetector();
 	bool Initialize();
-	bool Detect(double* x);
+	bool Detect(float* x);
 	void GetPiatch(PitchInfo& pitch);
 
 private:
 	const int _samplingRate;
 	const int _samplingSize;
 	std::shared_ptr<AutoCorrelation> _corr;
-	double* _r;
-	double* _m;
-	double* _x2;
-	double* _nsdf;
+	float* _r;
+	float* _m;
+	float* _x2;
+	float* _nsdf;
 	PitchInfo _pitch;
 
-	bool ComputeNsdf(double* x);
+	bool ComputeNsdf(float* x);
 	int AnalyzeNsdf();
 };
